Written by: Pascal Stumpf <pascal@openbsd.org>
OpenImageIO is a library for reading and writing images, and a bunch
of related classes, utilities, and applications.  There is a
particular emphasis on formats and functionality used in professional,
large-scale animation and visual effects work for film.  OpenImageIO
is used extensively in animation and VFX studios all over the world,
and is also incorporated into several commercial products.

ok ajacoutot@