Written by: Stuart Henderson <sthen@openbsd.org>
Drop separate OpenSSL flavour for FreeTDS. Change main package to using
OpenSSL rather than GnuTLS. Permit packages to be distributed. Bump shared
library versions to make sure they get updated.

As researched by jasper@, the GPL has a special exception when something
depends on software provided with the base OS, so it *is* ok for us to
distribute this.

This is preferred over using GnuTLS because GnuTLS pulls in a pthread
dependency; pthread overrides libc functions, so it is necessary
that a program using this is itself linked with pthread, causing a
problem if you want to use the FreeTDS library to connect to servers
using encryption from programs which aren't themselves linked with
pthread (e.g. /usr/sbin/httpd with mod_php, or perl with DBD::Sybase).

ok jasper@ ajacoutot@