Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import log4c-1.2.1.

Log4c is a library of C for flexible logging to files, syslog and 
other destinations. It is modeled after the Log for Java library 
(http://jakarta.apache.org/log4j/), staying as close to their API 
as is reasonable.

based on a submission from Markus Schatzl
ok jasper@