Written by: Marc Espie <espie@openbsd.org>
bite the bullet, reuse dpb's algorithm to make sure we get correct
PkgPath.

This implies restructuring the database code, and the start of a
better separation between variables handling, and actual insertion
in the database.

bump version to 2.0: this changes the paths that actually get recorded,
and pkglocatedb will need to be aware of that.