Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
If syslog-ng is started early during boot-up (that is if one replaces
the syslogd rc.d(8) script with the syslog_ng one), it will fail to
launch because ldconfig(8) hasn't had a chance to run yet. Add an
rc_pre() that will take care of that.

ok sthen@ steven@ (maintainer)