Written by: Ingo Schwarze <schwarze@openbsd.org>
In the footers of manuals formatted with groff,
do not print "OpenBSD 5.0" but "OpenBSD ports".
While here, remove some noise from the groff build log.
Bump the groff package.

Technically, this changes the contents of all packages that USE_GROFF,
but please refrain from bumping the world:  Having "OpenBSD 5.0" in
the footers of some ports manual until they are updated the next time,
or until the next libc bump if they aren't, is not a real problem.

string "OpenBSD ports" suggested and patch ok'ed by sthen@