Written by: Gonzalo L. Rodriguez <gonzalo@openbsd.org>
Update for Smplayer to 0.8.0 whit this features:

* A toolbar editor has been added. Now it's possible to select  the
buttons and controls that want to appear in the toolbars.
* New video filters: gradfun, blur and sharpen.
* Now it's possible to change the GUI (default, mini, mpc) at
runtime, no restart required.
* Sub files from opensubtitles should work again.
* Better support for chapters in video files.
* Bug fix: remote m3u files work from the favorites menu or command line.
* Internal changes in the single instance option (switch to
QtSingleApplication).

While here, GROFF is not needed, disable debug from edd@ and remove the
LIB_DEPENDS line, it's not needed since the x11/qt4 module is used.

Tested on i386 and ppc.

OK edd@, landry@ and aja@