Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
gtk-update-icon-cache:
Don't redirect errors to /dev/null and don't return true(1)
unconditionally. Instead, don't check for the existence of index.theme.
This will allow us to catch errors that may be happening because of a
missing dependency in the chain.

Some hidden issues may appear, in which case please contact me.

discussed with and ok blind jasper@