Written by: Marc Espie <espie@openbsd.org>
make very sure we don't get a wandering libtool
software has a piece of shit configure script (oxymoron) that does detect
libtool is present, but produces crap when it discovers it's not gnu libtool.
of course, checking for features is too complicated, so it just asks
libtool --version and checks for "GNU libtool".

morons.