Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import csync-0.50.6.

csync is a lightweight utility to synchronize files between two
directories on a system or between multiple systems.
<...>

inputs/ok jasper@

NOTE that currently the owncloud module is commented in the PLIST
because there is something wrong with our neon library for webdav which
segfaults. I am working on this -- for now at least local and sftp sync
work fine.