Written by: Stuart Henderson <sthen@openbsd.org>
import textproc/p5-LaTeX-Driver, merged from submissions by Andrew Fresh
and Chris Bennett.

The LaTeX::Driver module takes care of running and re-running latex on
a LaTeX document so that forward references, tables of contents, and
lists of figures and tables are resolved.  It will also run bibtex and
makeindex if it detects that a bibliography or in index have been
specified, and will re-run latex again one or more times until the
formatting of the document has stabilized.

landry@ ok to import (though slight tweaks since the version he saw).