Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Remove prelude. It's been unmaintained for years. Updating to a newer
version would require a lot of work and we have no maintainer; not to
mention the source code is not available right now.
A "security" software that is unmaintained is something that should live
in an attic.

ok jasper@ sthen@