Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
So, our linker does not hide private symbols... which results in a big
amount of crashes because gcr, libgnome-keyring and gnome-keyring share
some common (private) symbol names.
Miod says there is no workaround...

For gnome-shell, just stop linking to libgnome-keyring which we don't
need anyway since it is only needed by the network-manager functionnality
that is obviously disabled for us.