Written by: Stuart Henderson <sthen@openbsd.org>
import BIND 9.9.1-P1 as ports/net/isc-bind - this is newer than the
version of BIND than is in the base OS (some people require features
from this version e.g. DNS64), but note that it does not include
the hardening changes made to the version in base.

feedback from naddy@ giovanni@, ok giovanni@.

"BIND is open source software that implements the Domain Name System
(DNS) protocols for the Internet. It is a reference implementation
of those protocols, but it is also production-grade software,
suitable for use in high-volume and high-reliability applications."