Add a wip clang 3.2
This compiles and the compiler works.  Currently running a 3.2 compiled
kernel.  Package does not work yet due to PLIST issues that need to be
worked out.  I am sure there are a few more warts but have a look and
don't forget to bump the clang version to 3.2 in
lang/clang/clang.port.mk