Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import zurmo-0.7.10.

Zurmo is an open source CRM  application written in PHP utilizing
jQuery, Yii Framework, and RedBeanPHP.

The goal with Zurmo is to provide an easy-to-use, easy-to-customize CRM
application that can be adapted to any business use case. 
Special care has been taken to think through many different use cases
with a system designed to provide a high degree of flexibility, covering
a wide variety of use cases out of the box.
<...>

ok jasper@