Written by: Pascal Stumpf <pascal@openbsd.org>
Prepare emacs for PIE.  emacs2* currently use gmalloc, which is incompatible
with PIE.  We should be using the system malloc anyway for security reasons.
Unfortunately, there are yet more problems with PIE in emacs proper, so disable
it for now until a solution is found.

ok naddy@ sthen@ (on a previous version)