Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import qpdf-3.0.1.

QPDF is a command-line program that does structural, content-preserving
transformations on PDF files. It could have been called something like
pdf-to-pdf. It also provides many useful capabilities to developers of
PDF-producing software or for people who just want to look at the
innards of a PDF file to learn more about how they work.

This will be needed by an update to cups-filters which I am working on.
ok jasper@