Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
remove gtkaml; it served well as a vala testcase when there were no other
vala-dependant ports in the tree. now there are plenty of ports using vala
and this port keeps lagging behind and is a nightmare to keep fixing for
every vala release, so zap it.

ok non-rude aja@