Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import kyotocabinet-1.2.76.

Kyoto Cabinet is a library of routines for managing a database. The
database is a simple data file containing records, each is a pair of a
key and a value. Every key and value is serial bytes with variable
length. Both binary data and character string can be used as a key and a
value. Each key must be unique within a database. There is neither
concept of data tables nor data types. Records are organized in hash
table or B+ tree.

Available FLAVOR:
        optimized - build with optimizations for the local CPU


ok robert@ jasper@ sthen@