Written by: Matthias Kilian <kili@openbsd.org>
Conduits are an approach to the streaming data problem. It is meant
as an alternative to enumerators\/iterators, hoping to address the
same issues with different trade-offs based on real-world experience
with enumerators.

ok matthew, sthen.

(forgot to import this one the other day and imported devel/hs-lifted-base
twice instead)