Written by: Stuart Henderson <sthen@openbsd.org>
various tweaks to phpMyAdmin packaging.
feedback/OKs from aja@ william@ jasper@ giovanni@ kevlo@

- move MESSAGE to README.

- add missing dep on php-mysqli.

- change the unpack/fake-install method to use the standard
extract infrastructure.

- @sample contrib/htaccess into .htaccess so apache can use it by
default if allowoverride is enabled. this does some basic user-agent
checks to reduce the risk of phpmyadmin sites showing up in
search engines.

- provide an @sample'd apache config file which enables use of
the .htaccess file and restricts access to localhost. talk about
this in README so people know what's going on. (phpMyAdmin has
a history of security-related bugs, but because it's rather
commonly used it's much better to have it in ports so that
it can be updated easily; these changes make the default
installation safer).