Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import libvirt 0.10.1

Libvirt is collection of software that provides a convenient way to
manage virtual machines and other virtualization functionality, such as
storage and network interface management. These software pieces include
an API library, a daemon (libvirtd), and a command line utility (virsh).

Note that OpenBSD cannot act as a virtualization host yet, so libvirtd
is not provided in this package.
[...]

feedback/ok aja@