Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Be nice to cvs and manually add "core".

Add gstreamer1-0.11.93.
This is GStreamer, a framework for streaming media.
This package provides the core around which all other modules revolve.

This is a development release; version 1.0 is due somewhat soon and the
next big GNOME update will need it.
I tried to keep the dependencies as "less big" as possible to be able to
have a good out-of-the-box experience -- more obscure codecs can be
added later if there's a need for them.

ok jasper@