Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import glpi 0.83.4

GLPI is the Information Resource-Manager with an additional
Administration- Interface. You can use it to build up a database with an
inventory for your company (computer, software, printers...). It has
enhanced functions to make the daily life for the administrators easier,
like a job-tracking-system with mail-notification and methods to build a
database with basic information about your network-topology.

ok aja@