Written by: Stuart Henderson <sthen@openbsd.org>
Update nginx stable in ports to 1.2.x and add devel (1.3.x), this is
separate from the version in the base OS which deliberately does not include
the mail proxy and additional modules. rc.d/enginx script is added (not
rc.d/nginx to avoid conflicting with base). Lua module is also supported.
Most of the work done by william@, with a few tweaks by me and ok william.

README could use more work and as pointed out by ajacoutot we need to
review permissions of the tmp directory, but the diff is already long enough.
@pkgpath markers are set to update from the 1.0.x packages to 1.2.x (with or
without the passenger flavour as appropriate).