Written by: Stuart Henderson <sthen@openbsd.org>
SECURITY update to BIND 9.9.1-P3, fixes CVE-2012-4244
https://kb.isc.org/article/AA-00778

If a record with RDATA in excess of 65535 bytes is loaded into a
nameserver, a subsequent query for that record will cause named to exit
with an assertion failure.

This vulnerability can be exploited remotely against recursive servers
by inducing them to query for records provided by an authoritative
server. It affects authoritative servers if a zone containing this type
of resource record is loaded from file or provided via zone transfer.