Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
It's this time of the year again... update glib2 to the latest
available development release (glib2-2.33.14). A new stable is supposed
to come out within the next 2 weeks and we want to be proactive and get
some things in tree before the big GNOME 3.6 update that we will work on
at p2k12.

big thank to jasper@ for running this and the upcoming updates in a bulk
ok jasper@