Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
New support for sndio(7) from ratchov@ \o/

Initially as a patch for plugins-bad (where the sndio(7) module is
located upstream) -- but it makes no sense to install the whole kitchen
sink just to have sound -- so move it into plugins-base and adapt the
plumbing accordingly.

big thank you to ratchov@ for his work on this!