Written by: Jeremy Evans <jeremy@openbsd.org>
Now that ruby 1.9 is the default ruby implementation for gem/extconf
ports, for the ports that are built both on ruby 1.8 and ruby 1.9,
switch the category Makefiles to explicitly list the ruby18 FLAVOR
instead of the ruby19 FLAVOR.

Also, for home_run, fastri, and fastercsv, explicitly build only the
ruby 1.8 version of the port.  These libraries can run on ruby 1.9, but
it doesn't make sense to build a ruby 1.9 version by default.