Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Webkit is a rendering engine, so it makes no sense to force dependencies
on the whole gstreamer gang which are only uses by web browsers to
display audio/video.
So remove RUN_DEPENDS, it'll be explicitely added to webkit browsers
with a small comment so that maintainers of such browsers can decide
which gstreamer dependencies they need/want.

ok landry@ (maintainer)