Written by: Anthony J. Bentley <bentley@openbsd.org>
Import armagetronad-0.2.8.3.2.

Armagetron is a multiplayer game in 3d that attempts to emulate and
expand on the lightcycle sequence from the movie Tron. It's an old
school arcade game slung into the 21st century. Highlights include a
customizable playing arena, HUD, unique graphics, and AI bots. For the
more advanced player there are new game modes and a wide variety of
physics settings to tweak as well.

ok jeremy@ kirby@