Change the library order to be correct.
It should be -lm(*) -lpthread(*) -lc(*) _THEN_ -lclang_rt.<ARCH>

Because libc may depend on functions in libclang, however libclang
does _NOT_ depend on anything in libc. This is a difference between
libclang and libgcc, libgcc relied on functions in libc so it was
linked '-lgcc -lc -lgcc'. But this is not the case anymore.