Written by: David Coppa <dcoppa@openbsd.org>
Import x11/hs-dbus.

OK kili@

D-Bus is a simple, message-based protocol for inter-process
communication, which allows applications to interact with other
parts of the machine and the user's session using remote procedure
calls.
D-Bus is an essential part of the modern UNIX desktop, where it
replaces earlier protocols such as CORBA and DCOP.
This library is an implementation of the D-Bus protocol in Haskell.
It can be used to add D-Bus support to Haskell applications, without
the awkward interfaces common to foreign bindings.