Import blobby-1.0rc3.
Blobby Volley is a game featuring blobs playing volleyball.

Blobby Volley is one of the most popular freeware games. This is due
both to the simple gameplay, and funny design of the game characters.
Due to the relative quickness of the game it is often used as a game for
"in between" times.


ok ajacoutot@

Written by: Anthony J. Bentley <bentley@openbsd.org>