import swiggle 0.4 from openbsd-wip
swiggle is a small command line tool that generates HTML pages,
including thumbnail indexes, for given images (a so called
"web gallery"). It's intended to be easy to use, and since
it is written in C, it's quite speedy.

with feedback from ajacoutot@ and sthen@
ok sthen@ and Sergey Bronnikov (MAINTAINER

Written by: rpe <rpe@cvs.openbsd.org>