Import xournal-0.4.7.
Xournal is an application for notetaking, sketching, keeping a journal
using a stylus. It is similar to Microsoft Windows Journal or to other
alternatives such as Jarnal, Gournal, and NoteLab.

It can also be used to annotate PDF files (highlight, underline etc.),
and either save the annotations in a seperate file, or export the
annotated document as a new PDF file.


ok edd@

Written by: Anthony J. Bentley <bentley@openbsd.org>