Switch to OpenBSD method of finding needed libraries.
Instead of searching the host ld.so hints, allow -L paths to
act as -rpath-link so that needed libraries in the same path as
loaded libraries will be accepted with no special flags.

This reverts the previous change and bumps revision again.

NOTE: This behavior varies away from the standard GNU way of
building libraries and chooses the way OpenBSD has been performing
this operation since 2002.

After talking with marco@ and oga@ this was chosen to more closely
model the openbsd way of doing things since ports could be affected.